+++ index.js
@@ -6,7 +6,7 @@ const bodyParser = require('body-parser');
 const { v4: uuidv4 } = require('uuid');
 const session = require('express-session')
  
-const router = require('./router/auth.ejs');
+const router = require('./router/auth');
 
 //create the server
 const app = express();
@@ -38,6 +38,14 @@ app.use(session({
 
 //specify the routers
 app.use('/', router);
+
+app.get('/calendar', (req, res) =>{
+    res.render("calendar");
+})
+
+app.get("/signup", (req, res,) => {
+    res.render("signup");
+})
 
 //listen to the server
 app.listen(port, ()=>{
