+++ public/calendar-script.js
@@ -1,5 +1,4 @@
 // JS for calendar functions
-// JS for calendar functions
 const calendar = document.querySelector('.calendar'),
     date = document.querySelector('.date'),
     daysContainer = document.querySelector('.days'),
@@ -537,32 +536,6 @@ function getEvents() {
         return;
     }
 
-        eventsArr.push(...JSON.parse(localStorage.getItem('events')));
+        eventsArr.push(JSON.parse(localStorage.getItem('events')));
     }
 
-
-// script for the dashboard
-const body = document.querySelector('body'),
-    sidebar = body.querySelector('.sidebar'),
-    toggle = body.querySelector('.toggle'),
-    searchBtn = body.querySelector('.search-box'),
-    modeSwitch = body.querySelector('.toggle-switch'),
-    modeText = body.querySelector('.mode-text');
-
-toggle.addEventListener('click', () => {
-    sidebar.classList.toggle('close');
-});
-
-searchBtn.addEventListener('click', () => {
-    sidebar.classList.remove('close');
-});
-
-modeSwitch.addEventListener('click', () => {
-    body.classList.toggle('dark');
-
-    if(body.classList.contains('dark')){
-        modeText.innerText = 'Light Mode';
-    }else{
-        modeText.innerText = 'Dark Mode';
-    }
-});
