+++ public/style.css
@@ -706,24 +706,8 @@ body.dark .switch::before {
 /* start of calendar css */
 /* G-font: Poppins */
 
-:root {
-    --primary-clr: #b38add;
-}
 
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: 'Poppins', sans-serif;
-}
 
-body {
-    min-height: 100vh;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: #e2e1dc;
-}
 
 .container-calendar {
     position: relative;
