+++ router/auth.js
@@ -57,7 +57,7 @@ router.get('/dashboard',(req, res)=>{
     if(req.session.user){
         res.render('dashboard',{title:'Dashboard', user: req.session.user });
     }else{
-        res.send(403);
+        res.sendStatus(403);
     }
 });
 
@@ -72,5 +72,7 @@ router.get('/logout', (req, res)=>{
         }
     })
 });
+
+
 
 module.exports = router;
