+++ views/dashboard.ejs
@@ -29,7 +29,7 @@
                         </a>
                     </li>
                     <li class="nav-link">
-                        <a href="#">
+                        <a href="/calendar">
                             <i class='bx bx-calendar icon' ></i>
                             <span class="text nav-text">View</span>
                         </a>
@@ -57,7 +57,7 @@
 
             <div class="bottom-content">
                 <li class="">
-                    <a href="#">
+                    <a href="/logout">
                         <i class='bx bx-log-out icon' ></i>
                         <span class="text nav-text">Logout</span>
                     </a>
