+++ views/footer.ejs
@@ -1,4 +1,6 @@
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"></script>
-<script src="/script.js"></script>
+<script src="/static/dashboard-script.js"></script>
+<script src="/static/calendar-script.js"></script>
+
 </body>
 </html>
