+++ views/login.ejs
@@ -6,7 +6,7 @@
                 <h1>Log in</h1>
 
                 <p class="desc">Don't have an account?
-                    <a href="#">Create account</a>
+                    <a href="/signup">Create account</a>
                 </p>
 
                 <div>
@@ -42,11 +42,11 @@
                 </div>
 
                 <div class="input-group">
-                    <input type="email" class="input" placeholder="Email" value="kodego@test.com" required>
+                    <input type="email" name="email" class="input" placeholder="Email" value="kodego@test.com" required>
                 </div>
 
                 <div class="input-group">
-                    <input type="password" class="input" placeholder="Password" value="12345678" required>
+                    <input type="password" name="password" class="input" placeholder="Password" value="12345678" required>
                 </div>
 
                 <button type="submit" class="btn">
